test(c): split test_retry.py integration -> integration/test_retry.py (#40)
Moved TestRetryPipelineIntegration (4 tests) — pipeline with
intentionally-broken curation to force retry-backoff behaviour.
Five unit classes (state-only + mock-heavy connection cleanup) stay
for Wave D.
